Refactor the implementation
* Moved both the repository and the route under the `Configuration` subfolder namespace
* Implemented the endpoint related to the route
* Fixed the route to pass `is_dismissed` from the request body and not from the URL
