+++ src/dashboard/user-interface/configuration/site-kit-configuration-dismissal-route.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Yoast\WP\SEO\Dashboard\User_Interface;
+namespace Yoast\WP\SEO\Dashboard\User_Interface\Configuration;
 
 use Exception;
 use WP_Error;
 use WP_REST_Request;
 use WP_REST_Response;
 use Yoast\WP\SEO\Conditionals\No_Conditionals;
-use Yoast\WP\SEO\Dashboard\Infrastructure\Permanently_Dismissed_Site_Kit_Configuration_Repository_Interface;
+use Yoast\WP\SEO\Dashboard\Infrastructure\Configuration\Permanently_Dismissed_Site_Kit_Configuration_Repository_Interface;
 use Yoast\WP\SEO\Main;
 use Yoast\WP\SEO\Routes\Route_Interface;
 
@@ -21,6 +21,13 @@ use Yoast\WP\SEO\Routes\Route_Interface;
 class Site_Kit_Configuration_Dismissal_Route implements Route_Interface {
 
 	use No_Conditionals;
+
+	/**
+	 *  The namespace for this route.
+	 *
+	 * @var string
+	 */
+	public const ROUTE_NAMESPACE = Main::API_V1_NAMESPACE;
 
 	/**
 	 *  The prefix for this route.
@@ -54,21 +61,13 @@ class Site_Kit_Configuration_Dismissal_Route implements Route_Interface {
 	 */
 	public function register_routes() {
 		\register_rest_route(
-			Main::API_V1_NAMESPACE,
+			self::ROUTE_NAMESPACE,
 			self::ROUTE_PREFIX,
 			[
 				[
 					'methods'             => 'POST',
 					'callback'            => [ $this, 'set_site_kit_configuration_permanent_dismissal' ],
 					'permission_callback' => [ $this, 'check_capabilities' ],
-					'args'                => [
-						'is_dismissed' => [
-							'required'          => true,
-							'type'              => 'bool',
-							'default'           => false,
-							'sanitize_callback' => 'rest_sanitize_boolean',
-						],
-					],
 				],
 			]
 		);
@@ -82,8 +81,7 @@ class Site_Kit_Configuration_Dismissal_Route implements Route_Interface {
 	 * @return WP_REST_Response|WP_Error The success or failure response.
 	 */
 	public function set_site_kit_configuration_permanent_dismissal( WP_REST_Request $request ) {
-		$params       = $request->get_params();
-		$is_dismissed = $params['is_dismissed'];
+		$is_dismissed = $request->get_param( 'is_dismissed' );
 
 		try {
 			$result = $this->permanently_dismissed_site_kit_configuration_repository->set_site_kit_configuration_dismissal( $is_dismissed );
