+++ tests/Unit/Dashboard/User_Interface/Configuration/Check_Capabilities_Test.php
@@ -1,15 +1,16 @@
 <?php
 
-namespace Yoast\WP\SEO\Tests\Unit\Dashboard\User_Interface;
+namespace Yoast\WP\SEO\Tests\Unit\Dashboard\User_Interface\Configuration;
 
 use Brain\Monkey\Functions;
+use function Brain\Monkey\Functions;
 
 /**
  * Abstract class for the Permanently Dismissed Site Kit Configuration Repository tests.
  *
  * @group site_kit_configuration_permanent_dismissal_route
  *
- * @covers Yoast\WP\SEO\Dashboard\User_Interface\Site_Kit_Configuration_Permanent_Dismissal_Route::check_capabilities
+ * @covers Yoast\WP\SEO\Dashboard\User_Interface\Configuration\Site_Kit_Configuration_Permanent_Dismissal_Route::check_capabilities
  *
  * @phpcs:disable Yoast.NamingConventions.ObjectNameDepth.MaxExceeded
  */
