+++ tests/Unit/Dashboard/User_Interface/Configuration/Set_Site_Kit_Permanent_Dismissal_Test.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yoast\WP\SEO\Tests\Unit\Dashboard\User_Interface;
+namespace Yoast\WP\SEO\Tests\Unit\Dashboard\User_Interface\Configuration;
 
 use Mockery;
 use WP_REST_Request;
@@ -11,7 +11,7 @@ use WP_REST_Response;
  *
  * @group site_kit_configuration_permanent_dismissal_route
  *
- * @covers Yoast\WP\SEO\Dashboard\User_Interface\Site_Kit_Configuration_Permanent_Dismissal_Route::set_site_kit_configuration_permanent_dismissal
+ * @covers Yoast\WP\SEO\Dashboard\User_Interface\Configuration\Site_Kit_Configuration_Permanent_Dismissal_Route::set_site_kit_configuration_permanent_dismissal
  *
  * @phpcs:disable Yoast.NamingConventions.ObjectNameDepth.MaxExceeded
  */
@@ -45,13 +45,9 @@ final class Set_Site_Kit_Permanent_Dismissal_Test extends Abstract_Site_Kit_Conf
 
 		$wp_rest_request = Mockery::mock( WP_REST_Request::class );
 		$wp_rest_request
-			->expects( 'get_params' )
+			->expects( 'get_param' )
 			->once()
-			->andReturn(
-				[
-					'is_dismissed'         => $is_dismissed,
-				]
-			);
+			->andReturn( $is_dismissed );
 
 		$this->assertInstanceOf(
 			'WP_REST_Response',
